intake-mcp/mcp-client/requirements.txt
fastmcp
google-genai
httpx